removed node* tail var
I could not figure out the tail issue so I just removed it. The list is doubly linked so I can just navigate to the end of the list if I wish (although it will be 0[n] ). If this becomes a problem later, I will look back into adding a node* tail var